Changes in the double vilt architecture
I corrected some mistakes in the architecture in order to do what it was supposed to do. Then I added the option to load my pretrained weights to the preprocess part of the architecture. In the end I changed the default option of fine_tune_all = False to not fine tune only the final model (this maybe I'll change it back to how it was again).