Transformer Decoder
I implemented the concept of multiple multihead attention through transformer decoders. The idea is that we have multiple layers (2 for now) of image guided attention on the question and question guided attention to the image and then these better representations go through a final multi head attention module to calculate the attention scores. The rest of the process remains the same as in the original double_vilt architecture