First ViT_ViLT architecture
I did the basic modifications for the ViT_ViLT model to be able to be trained. That includes modifying models.py to create the model class and train.py to incorporate it in the training pipeline. Also it was necessary to modify utility.py and isvqa.py in order to create a new dataset class that will prepare the input images in 224x224 size, which is compatible with ViT. More refinements are needed, which will be added in the future.